Fixed bugs and improved display of download progress. Fixes #70

+++ filedownloader.h
@@ -68,7 +68,7 @@ private:
     utimbuf times;
 
     QString correctUnit(qint64 bytes);
-    qint64 correctSize(qint64 bytes);
+    double correctSize(qint64 bytes);
 
     QTime downloadTime;
